+++ nemo/collections/llm/configs/__init__.py
@@ -1,0 +1,13 @@
+from nemo.collections.llm.configs import llama2_7b, llama3_8b, llama3_8b_16k, llama3_8b_64k, mistral
+from nemo.collections.llm.configs.log.default import default_log
+from nemo.collections.llm.configs.optim import adam
+
+__all__ = [
+    "llama3_8b",
+    "llama3_8b_16k",
+    "llama3_8b_64k",
+    "llama2_7b",
+    "mistral",
+    "adam",
+    "default_log",
+]
