+++ nemo/collections/llm/configs/llama3_8b_16k.py
@@ -4,8 +4,8 @@ from nemo import lightning as nl
 from nemo.collections.llm.api import pretrain
 from nemo.collections.llm.gpt.data.api import squad
 from nemo.collections.llm.gpt.model.llama import Llama3Config8B, LlamaModel
-from nemo.collections.llm.models.log.default import default_log
-from nemo.collections.llm.models.optim.adam import adam_with_cosine_annealing
+from nemo.collections.llm.configs.log.default import default_log
+from nemo.collections.llm.configs.optim.adam import adam_with_cosine_annealing
 from nemo.collections.llm.utils import Partial, factory
 
 NAME = "llama3_8b_16k"
